refactor: 유효성 검사 함수 한 곳에 모아놔라
