+++ src/app/asset/validation/checkChargeAmount.js
@@ -1,4 +1,4 @@
-import { createCheckNumber } from './index.js';
+import { createCheckNumber } from './common.js';
 import { ERROR_MSG } from '../constants/index.js';
 
 const checkProductQuantity = createCheckNumber(ERROR_MSG.requireChargeCoin);
