+++ src/app/asset/validation/checkProductName.js
@@ -1,8 +1,10 @@
-import { createCheckEveryFunction, createValidFunction, isRequired } from './index.js';
+import { createCheckEveryFunction, createValidFunction, isRequired } from './common.js';
 import { ERROR_MSG } from '../constants/index.js';
 import { isUniqueProductName } from '../../localStorage/inventory.js';
 
-export default createCheckEveryFunction([
+const checkProductName = createCheckEveryFunction([
     createValidFunction(isRequired, ERROR_MSG.requireProductName),
     createValidFunction(isUniqueProductName, ERROR_MSG.overlapProductName),
 ]);
+
+export default checkProductName;
